@knizhnik 2022-09-26
Looks like my assumption concerning where bottleneck is was wrong.
At least for my `test_layer_map.py` test.
I assume that we have a lot of layers with different LSN belonging to the same key range. as as far as currently 'LayerMap::search` uses whole LSN range (starting from zero), then R-Tree lookup for intersecting layer will return a lot of candidates which we have to iterate in `LayerMap::`search`. But is is not true for `test_layer_map.py`.  Because of small  `checkpoint_distance=8192` it produce a lot of small layers, but with different key range. So average number of traverse layers s just 3. So  this is why my optimization with cutting-off nodes with smaller LSN will not help much in this case.

Attached please find FameGraph with query execution in  `test_layer_map.py`.  As you can see, LayerMap::`search` is taken only 15% - sightly more than disk_btree search (12%).
![layer_map](https://user-images.githubusercontent.com/156922/192217117-c753ba4a-df7e-406e-81a7-99711f40cd0e.svg)