@knizhnik 2022-09-26
But with compaction we have different story:
![compaciton](https://user-images.githubusercontent.com/156922/192220904-128b8053-a844-4472-9d1d-36d319214c67.svg)
Here we actually traverse ALL layers. But it is not a problem of index (R-Tree), but our LayerMap code.