@knizhnik 2022-11-30
> Do you have new performance profiles for these changes?

In theory such traversal can reduce number of tree nodes (layers) which has to be inspected.
Practically I have tested this PR on three different workloads: one artificial (test_layer_map) and two read (grabbed Kontor layers layout). 

I tried to understand why so much time is spent in layer traversal during compaction in production.
The only real operation I can do without having content of real layers is `time_for_new_image_layer` which is called by   `create_image_layers` to check whether new image layer need to be created. The most of time in `time_for_new_image_layer`  is spent in `LayerMap::image_coverage` which in turn calls `find_latest_image`. So average number of traversed layers with this patch is 15 comparing with 63 for main. But it doesn't have large impact on operation time, 278/432/310 msec vs 255/636/288/msec (min/max/avg). But as far as time is in any case very small, most likely it is not the critical path. 

Results of other benchmarks are more impressive (especially on @bojanserafimov  bench_layer_map  which performs `LayerMap::search` on real layer map):

| test | main | this PR |
| ----- | ------ | ---------- |
| test_layer_map | 8.8827 sec | 6.551 sec | 
| cargo bench --bench bench_layer_map |  203.75 ms | 74.431 ms |