@knizhnik 2022-11-30
{{{
-   89,59%     0,17%  background op w  pageserver         [.] pageserver::tenant::timeline::Timeline::create_image_layers                                                                  
   - 89,42% pageserver::tenant::timeline::Timeline::create_image_layers                                                                                                                    
      - 87,71% pageserver::tenant::layer_map::LayerMap::image_coverage                                                                                                                     
         - 53,31% pageserver::tenant::layer_map::LayerMap::find_latest_layer                                                                                                               
            + 44,10% pageserver::tenant::layer_map::LayerMap::find_latest_layer_recurs                                                                                                     
            + 9,03% tracing_core::event::Event::dispatch                                                                                                                                   
         - 29,85% <rstar::algorithm::iterators::SelectionIterator<T,Func> as core::iter::traits::iterator::Iterator>::next                                                                 
              6,91% rstar::point::PointExt::all_component_wise                                                                                                                             
              4,05% rstar::point::PointExt::all_component_wise                                                                                                                             
              3,19% <smallvec::SmallVec<A> as core::iter::traits::collect::Extend<<A as smallvec::Array>::Item>>::extend                                                                   
           1,02% <pageserver::tenant::image_layer::ImageLayer as pageserver::tenant::storage_layer::Layer>::get_lsn_range                                                                  
           0,90% core::ops::range::Range<Idx>::contains                                                                                                                                    
           0,79% <pageserver::tenant::delta_layer::DeltaLayer as pageserver::tenant::storage_layer::Layer>::get_lsn_range                                                                  
      + 1,66% pageserver::tenant::layer_map::LayerMap::count_deltas                                                                                                                        
}}}