@knizhnik 2022-12-01
More results for synthetic layer map test added by @bayandin (https://github.com/neondatabase/neon/pull/2979):

| branch | total time | query time |
| --------- | ------------ | -------------- |
| main    | 3m45s     |138.55 ms  |
| this PR| 3m15s     | 242.47 ms |

So when layer map tree is so wide (lrgr key dimension) my  LSN-descending traversal order becomes less efficient.