@knizhnik 2022-12-01
> > More results for synthetic layer map test added by @bayandin (#2979)
> 
> By @bojanserafimov I guess wink

Yes, sorry - looks like my brain is using LR-1 parser:)